python_code/requirements.txt
joblib==1.3.2
numpy==1.23.5
pandas==2.0.3
sklearn_genetic==0.5.1
scikit_learn==1.2.0
scipy==1.11.2
skops==0.8.0
statsmodels==0.14.0
tsfresh==0.20.1
